requirements.txt
opencv-python
pytesseract
pillow
requests
huggingface_hub
pyautogui
webbrowser
